fix: Support empty arrays
strings.Split returns an array with the initial string if the separator
cannot be found. This means that if the array on the database is empty
`{}`, this will try to convert an empty string to a number.

Avoid this by checking if the scanned string is empty after removing the
brackets, and setting an empty array in that case.